Fix error in listing volumes if volume retention is large
There was a bug with integer out of range for big volume retention value.

There has been also fixed casting the following Job and Media values:

 * Media.FirstWritten
 * Media.LastWritten
 * Job.SchedTime
 * Job.StartTime
 * Job.EndTime
 * Job.RealEndTime

Otherwise in January 2038 they could reach out of range as well.

This bug has been reported and checked by @alfonso-escribano in Bacularis API repository:

https://github.com/bacularis/bacularis-api/issues/9

+++ API/Modules/JobManager.php
@@ -136,10 +136,10 @@ LEFT JOIN FileSet ON FileSet.FilesetId=Job.FilesetId '
 		$db_params = $this->getModule('api_config')->getConfig('db');
 		if ($db_params['type'] === Database::PGSQL_TYPE) {
 			$add_cols .= '
-				CAST(EXTRACT(EPOCH FROM Job.SchedTime) AS INTEGER) AS schedtime_epoch,
-				CAST(EXTRACT(EPOCH FROM Job.StartTime) AS INTEGER) AS starttime_epoch,
-				CAST(EXTRACT(EPOCH FROM Job.EndTime) AS INTEGER) AS endtime_epoch,
-				CAST(EXTRACT(EPOCH FROM Job.RealEndTime) AS INTEGER) AS realendtime_epoch
+				CAST(EXTRACT(EPOCH FROM Job.SchedTime) AS BIGINT) AS schedtime_epoch,
+				CAST(EXTRACT(EPOCH FROM Job.StartTime) AS BIGINT) AS starttime_epoch,
+				CAST(EXTRACT(EPOCH FROM Job.EndTime) AS BIGINT) AS endtime_epoch,
+				CAST(EXTRACT(EPOCH FROM Job.RealEndTime) AS BIGINT) AS realendtime_epoch
 			';
 		} elseif ($db_params['type'] === Database::MYSQL_TYPE) {
 			$add_cols .= "
